restructures owned object code
