+++ explorer/client/src/components/ownedobjects/views/OwnedNFTView.tsx
@@ -1,6 +1,22 @@
+// Copyright (c) 2022, Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import { trimStdLibPrefix, truncate } from '../../../utils/stringUtils';
+import DisplayBox from '../../displaybox/DisplayBox';
+import Longtext from '../../longtext/Longtext';
+import { type DataType } from '../OwnedObjectConstants';
+
+import type BN from 'bn.js';
+
 import styles from './OwnedObjects.module.css';
 
-export function OwnedNFTView({ results }: { results: resultType }) {
+export default function OwnedNFTView({ results }: { results: DataType }) {
+    const alttextgen = (value: number | string | boolean | BN): string =>
+        truncate(String(value), 19);
+
+    const lastRowHas2Elements = (itemList: any[]): boolean =>
+        itemList.length % 3 === 2;
+
     return (
         <div id="ownedObjects" className={styles.ownedobjects}>
             {results.map((entryObj, index1) => (
